Bump 3.16.0: drop split-button, project + new-project actions use a single dropdown trigger
ActionSplitButton (left primary + right caret) is replaced by ActionDropdownButton — one button that opens the menu on click. Row 1 of the menu is the built-in default (open New project modal / 'work on <slug>'); rows 2+ are user-configured projectActions / newProjectActions. Drops the split-only CSS overrides (.row-action.work-on, .modal-button.work-on-button, .new-project-button wrapper rules) made redundant by the single-button shape. Docs paragraphs for terminal.projectActions / terminal.newProjectActions rewritten to match the new shape.
