+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "condash",
-  "version": "3.15.1",
+  "version": "3.16.0",
   "description": "Markdown project dashboard for the conception tree (Electron build)",
   "private": true,
   "main": "dist-electron/main/index.js",
