--- src/renderer/action-split-button.css
@@ -1,221 +1,0 @@
-/* Split button: primary action + caret dropdown.
- * Used on project cards, project preview modal, and the "+ New project" button. */
-
-.action-split-button {
-  display: inline-flex;
-  align-items: stretch;
-  border: 1px solid var(--border);
-  border-radius: 6px;
-  background: var(--bg-subtle);
-  overflow: hidden;
-  /* No transition / no hover on the wrapper itself — each inner button
-   * owns its own hover. With the wrapper also reacting to hover, the
-   * primary and caret looked like one merged button when the user
-   * hovered either half. */
-}
-
-/* Hover each inner button INDEPENDENTLY so the user sees which half
- * they're about to fire. The half they're NOT on stays at the wrapper's
- * resting bg-subtle — the contrast across the divider is the visual
- * cue. color-mix amps the hover background a notch over bg-hover so
- * the two halves don't look identical in dark mode. */
-.action-split-button .action-split-primary:hover,
-.action-split-button .action-split-caret:hover {
-  background: color-mix(in srgb, var(--accent) 14%, var(--bg-hover));
-  color: var(--accent);
-}
-
-.action-split-primary {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  gap: 6px;
-  padding: 6px 8px;
-  border: none;
-  background: transparent;
-  color: var(--text);
-  cursor: pointer;
-  font: inherit;
-  font-size: var(--text-md);
-  font-weight: 600;
-  transition: color 100ms ease;
-}
-
-.action-split-caret {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: 22px;
-  padding: 0;
-  border: none;
-  border-left: 1px solid var(--border);
-  background: transparent;
-  color: var(--text);
-  cursor: pointer;
-  font: inherit;
-  transition:
-    color 100ms ease,
-    border-color 100ms ease;
-}
-
-.action-split-caret svg {
-  width: 12px;
-  height: 12px;
-}
-
-/* Menu portal'd below the trigger. width: max-content + min/max so the
- * menu grows to fit the widest row instead of letting the label/badge
- * ellipsis kick in at 260 px. */
-.action-split-menu {
-  width: max-content;
-  min-width: 260px;
-  max-width: 520px;
-  background: var(--bg);
-  border: 1px solid var(--border);
-  border-radius: 8px;
-  box-shadow: 0 8px 24px rgba(0, 0, 0, 0.18);
-  padding: 4px;
-  display: flex;
-  flex-direction: column;
-  gap: 1px;
-}
-
-.action-split-menu-item {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 6px 10px;
-  border: none;
-  border-radius: 6px;
-  background: transparent;
-  color: var(--text);
-  cursor: pointer;
-  font: inherit;
-  font-size: var(--text-sm);
-  text-align: left;
-  transition: background 80ms ease;
-}
-
-.action-split-menu-item:hover {
-  background: var(--bg-hover);
-}
-
-.action-split-menu-default {
-  font-weight: 500;
-}
-
-.action-split-menu-divider {
-  height: 1px;
-  margin: 2px 6px;
-  background: var(--border);
-}
-
-.action-split-menu-icon svg,
-.action-split-menu-bullet {
-  width: 14px;
-  height: 14px;
-  flex-shrink: 0;
-  color: var(--text-faint);
-}
-
-.action-split-menu-label {
-  flex: 1 1 auto;
-  min-width: 0;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.action-split-menu-submit {
-  color: var(--text-faint);
-  font-size: var(--text-2xs);
-  margin-left: auto;
-  flex-shrink: 0;
-}
-
-.action-split-menu-badge {
-  color: var(--text-faint);
-  font-size: var(--text-2xs);
-  margin-left: auto;
-  flex-shrink: 0;
-  text-transform: uppercase;
-  letter-spacing: 0.06em;
-}
-
-/* Inline launcher hint on configured action items — surfaces the bound
- * launcher right next to the action so users see at a glance which term
- * mode the entry will spawn into. Sits between the label (which ellipses
- * if long) and the trailing submit glyph. */
-.action-split-menu-launcher {
-  color: var(--accent);
-  font-size: var(--text-2xs);
-  font-family: var(--font-mono, ui-monospace, Menlo, monospace);
-  margin-left: auto;
-  padding: 1px 6px;
-  border: 1px solid var(--accent);
-  border-radius: 4px;
-  flex-shrink: 0;
-  opacity: 0.85;
-}
-
-.action-split-menu-launcher + .action-split-menu-submit {
-  margin-left: 6px;
-}
-
-/* Card-specific sizing override. The selectors are *compound* (no space)
- * because `row-action` and `work-on` live on the same element as
- * `action-split-button` — a descendant combinator would never match. */
-.action-split-button.row-action.work-on {
-  /* `.row .row-action` in projects-pane.css forces a hard 28×28 box on every
-   * row action button. That sizing fits a single-icon button but squeezes
-   * this two-child split button so the caret either renders at 0 px or is
-   * clipped by `overflow: hidden`. Reset to natural width / height + drop
-   * the row-action chrome we don't want here (the inner buttons supply
-   * their own background and padding). */
-  width: auto;
-  height: auto;
-  border: none;
-  background: transparent;
-  padding: 0;
-  overflow: visible;
-}
-
-.action-split-button.row-action.work-on:hover {
-  background: transparent;
-  border-color: transparent;
-}
-
-.action-split-button.row-action.work-on .action-split-primary {
-  padding: 4px;
-}
-
-.action-split-button.row-action.work-on .action-split-caret {
-  width: 18px;
-  border-left: none;
-}
-
-/* Modal-header sizing override — same compound-selector fix as above. */
-.action-split-button.modal-button.work-on-button {
-  width: auto;
-  height: auto;
-  border: none;
-  background: transparent;
-  padding: 0;
-  overflow: visible;
-}
-
-.action-split-button.modal-button.work-on-button .action-split-primary {
-  padding: 4px;
-}
-
-.action-split-button.modal-button.work-on-button .action-split-caret {
-  width: 20px;
-  border-left: none;
-}
-
-/* "+ New project" header action — pad the caret enough that the chevron
- * gets a comfortable 24 px hit target instead of the default 22 px so the
- * dropdown isn't a pixel-hunt next to the primary text. */
-.action-split-button.new-project-button .action-split-caret {
-  width: 24px;
-}
